make dbCAN substrate output optional

+++ modules/nf-core/rundbcan/cazymeannotation/main.nf
@@ -14,7 +14,7 @@ process RUNDBCAN_CAZYMEANNOTATION {
     output:
     tuple val(meta), path("${prefix}_overview.tsv")            , emit: cazyme_annotation
     tuple val(meta), path("${prefix}_dbCAN_hmm_results.tsv")   , emit: dbcanhmm_results
-    tuple val(meta), path("${prefix}_dbCANsub_hmm_results.tsv"), emit: dbcansub_results
+    tuple val(meta), path("${prefix}_dbCANsub_hmm_results.tsv"), emit: dbcansub_results, optional: true
     tuple val(meta), path("${prefix}_diamond.out")             , emit: dbcandiamond_results
     path  "versions.yml"                                       , emit: versions
 
@@ -34,7 +34,9 @@ process RUNDBCAN_CAZYMEANNOTATION {
 
     mv overview.tsv ${prefix}_overview.tsv
     mv dbCAN_hmm_results.tsv ${prefix}_dbCAN_hmm_results.tsv
-    mv dbCANsub_hmm_results.tsv ${prefix}_dbCANsub_hmm_results.tsv
+    if [ -f dbCANsub_hmm_results.tsv ]; then
+        mv dbCANsub_hmm_results.tsv ${prefix}_dbCANsub_hmm_results.tsv
+    fi
     mv diamond.out ${prefix}_diamond.out
 
     cat <<-END_VERSIONS > versions.yml
